Capitalize all form labels
* Currently we capitalize all the form fields using a
simple_form initializer. However, this is throwing some errors in edge
cases.
* Wrapped this in a form and http://www.w3.org/TR/selectors/#descendant-combinators
after talking with @kaytcampbell. This will make sure that we create all
form elements inside of an actual form element. All new styling for this
file should be nested under form * too (we'll move the rest over as we
go)